Taini files compatibility
The main program (coherence.py) offers the possibility to load taini files of 14 eeg electrodes instead of the openephys recordings of 32 EEG electrodes. 

It works well until it tries to downsample mixing the brain states with the voltages. I'll wait until we have more data to compare to code a way of solving it.