Coherence analysis working with npy files
Once done the downsampling, filtering and split by brain state in a previous process, the coherence analysis will work just with .npy files of a single brain state. 

The coherence analysis still joins the recordings from the same animal in a single recording before running the analysis.

That will save both time and a lot of memory.